chore: Add HttpJson Unary Callable Showcase Test
